requirements.txt
Django>=4.2
gunicorn
reportlab
WeasyPrint>=60.0
docxtpl
python-docx
Pillow
